Unifies remote provider service into integrations
